@@ -27,7 +27,7 @@ class AudioCache {
   ///
   /// If not set, the AudioCache will create and return a new instance of AudioPlayer every call, allowing for simultaneous calls.
   /// If this is set, every call will overwrite previous calls.
-  AudioPlayer fixedPlayer;
+  AudioPlayer? fixedPlayer;
 
   /// This flag should be set to true, if player is used for playing internal notifications
   ///
@@ -96,7 +96,7 @@ class AudioCache {
     if (!loadedFiles.containsKey(fileName)) {
       loadedFiles[fileName] = await fetchToMemory(fileName);
     }
-    return loadedFiles[fileName];
+    return loadedFiles[fileName]!;
   }
 
   AudioPlayer _player(PlayerMode mode) {
@@ -113,11 +113,11 @@ class AudioCache {
   Future<AudioPlayer> play(
     String fileName, {
     double volume = 1.0,
-    bool isNotification,
+    bool? isNotification,
     PlayerMode mode = PlayerMode.MEDIA_PLAYER,
     bool stayAwake = false,
     bool recordingActive = false,
-    bool duckAudio,
+    bool? duckAudio,
   }) async {
     String url = await getAbsoluteUrl(fileName);
     AudioPlayer player = _player(mode);
@@ -139,11 +139,11 @@ class AudioCache {
   Future<AudioPlayer> playBytes(
     Uint8List fileBytes, {
     double volume = 1.0,
-    bool isNotification,
+    bool? isNotification,
     PlayerMode mode = PlayerMode.MEDIA_PLAYER,
     bool loop = false,
-    bool stayAwake,
-    bool recordingActive,
+    bool stayAwake = false,
+    bool recordingActive = false,
   }) async {
     AudioPlayer player = _player(mode);
 
@@ -169,7 +169,7 @@ class AudioCache {
   Future<AudioPlayer> loop(
     String fileName, {
     double volume = 1.0,
-    bool isNotification,
+    bool? isNotification,
     PlayerMode mode = PlayerMode.MEDIA_PLAYER,
     bool stayAwake = false,
   }) async {
