@@ -96,25 +96,25 @@ void _backgroundCallbackDispatcher() {
   WidgetsFlutterBinding.ensureInitialized();
 
   // Reference to the onAudioChangeBackgroundEvent callback.
-  Function(AudioPlayerState) onAudioChangeBackgroundEvent;
+  Function(AudioPlayerState)? onAudioChangeBackgroundEvent;
 
   // This is where the magic happens and we handle background events from the
   // native portion of the plugin. Here we message the audio notification data
   // which we then pass to the provided callback.
   _channel.setMethodCallHandler((MethodCall call) async {
-    Function _performCallbackLookup() {
+    Function(AudioPlayerState) _performCallbackLookup() {
       final CallbackHandle handle = CallbackHandle.fromRawHandle(
           call.arguments['updateHandleMonitorKey']);
 
       // PluginUtilities.getCallbackFromHandle performs a lookup based on the
       // handle we retrieved earlier.
-      final Function closure = PluginUtilities.getCallbackFromHandle(handle);
+      final Function? closure = PluginUtilities.getCallbackFromHandle(handle);
 
       if (closure == null) {
         print('Fatal Error: Callback lookup failed!');
         // exit(-1);
       }
-      return closure;
+      return closure as Function(AudioPlayerState);
     }
 
     final Map<dynamic, dynamic> callArgs = call.arguments as Map;
@@ -122,11 +122,11 @@ void _backgroundCallbackDispatcher() {
       onAudioChangeBackgroundEvent ??= _performCallbackLookup();
       final String playerState = callArgs['value'];
       if (playerState == 'playing') {
-        onAudioChangeBackgroundEvent(AudioPlayerState.PLAYING);
+        onAudioChangeBackgroundEvent!(AudioPlayerState.PLAYING);
       } else if (playerState == 'paused') {
-        onAudioChangeBackgroundEvent(AudioPlayerState.PAUSED);
+        onAudioChangeBackgroundEvent!(AudioPlayerState.PAUSED);
       } else if (playerState == 'completed') {
-        onAudioChangeBackgroundEvent(AudioPlayerState.COMPLETED);
+        onAudioChangeBackgroundEvent!(AudioPlayerState.COMPLETED);
       }
     } else {
       assert(false, "No handler defined for method type: '${call.method}'");
@@ -182,7 +182,7 @@ class AudioPlayer {
   /// Enables more verbose logging.
   static bool logEnabled = false;
 
-  AudioPlayerState _audioPlayerState;
+  late AudioPlayerState _audioPlayerState;
 
   AudioPlayerState get state => _audioPlayerState;
 
@@ -249,7 +249,7 @@ class AudioPlayer {
 
   /// Handler of changes on player state.
   @deprecated
-  AudioPlayerStateChangeHandler audioPlayerStateChangeHandler;
+  AudioPlayerStateChangeHandler? audioPlayerStateChangeHandler;
 
   /// Handler of changes on player position.
   ///
@@ -260,7 +260,7 @@ class AudioPlayer {
   ///
   /// This is deprecated. Use [onAudioPositionChanged] instead.
   @deprecated
-  TimeChangeHandler positionHandler;
+  TimeChangeHandler? positionHandler;
 
   /// Handler of changes on audio duration.
   ///
@@ -269,7 +269,7 @@ class AudioPlayer {
   ///
   /// This is deprecated. Use [onDurationChanged] instead.
   @deprecated
-  TimeChangeHandler durationHandler;
+  TimeChangeHandler? durationHandler;
 
   /// Handler of player completions.
   ///
@@ -280,7 +280,7 @@ class AudioPlayer {
   ///
   /// This is deprecated. Use [onPlayerCompletion] instead.
   @deprecated
-  VoidCallback completionHandler;
+  VoidCallback? completionHandler;
 
   /// Handler of seek completion.
   ///
@@ -288,7 +288,7 @@ class AudioPlayer {
   ///
   /// This is deprecated. Use [onSeekComplete] instead.
   @deprecated
-  SeekHandler seekCompleteHandler;
+  SeekHandler? seekCompleteHandler;
 
   /// Handler of player errors.
   ///
@@ -296,7 +296,7 @@ class AudioPlayer {
   ///
   /// This is deprecated. Use [onPlayerError] instead.
   @deprecated
-  ErrorHandler errorHandler;
+  ErrorHandler? errorHandler;
 
   /// An unique ID generated for this instance of [AudioPlayer].
   ///
@@ -308,30 +308,27 @@ class AudioPlayer {
   PlayerMode mode;
 
   /// Creates a new instance and assigns an unique id to it.
-  AudioPlayer({this.mode = PlayerMode.MEDIA_PLAYER, this.playerId}) {
-    this.mode ??= PlayerMode.MEDIA_PLAYER;
-    this.playerId ??= _uuid.v4();
+  AudioPlayer({this.mode = PlayerMode.MEDIA_PLAYER, this.playerId = ''}) {
+    this.playerId = this.playerId == '' ? _uuid.v4() : this.playerId;
     players[playerId] = this;
 
     if (defaultTargetPlatform == TargetPlatform.iOS) {
       // Start the headless audio service. The parameter here is a handle to
       // a callback managed by the Flutter engine, which allows for us to pass
       // references to our callbacks between isolates.
-      final CallbackHandle handle =
+      final CallbackHandle? handle =
           PluginUtilities.getCallbackHandle(_backgroundCallbackDispatcher);
       assert(handle != null, 'Unable to lookup callback.');
       _invokeMethod('startHeadlessService', {
-        'handleKey': <dynamic>[handle.toRawHandle()],
+        'handleKey': <dynamic>[handle?.toRawHandle()],
       });
     }
   }
 
   Future<int> _invokeMethod(
     String method, [
-    Map<String, dynamic> arguments,
+    Map<String, dynamic> arguments = const {},
   ]) {
-    arguments ??= const {};
-
     final Map<String, dynamic> withPlayerId = Map.of(arguments)
       ..['playerId'] = playerId
       ..['mode'] = mode.toString();
@@ -350,11 +347,11 @@ class AudioPlayer {
     // Start the headless audio service. The parameter here is a handle to
     // a callback managed by the Flutter engine, which allows for us to pass
     // references to our callbacks between isolates.
-    final CallbackHandle handle =
+    final CallbackHandle? handle =
         PluginUtilities.getCallbackHandle(_backgroundCallbackDispatcher);
     assert(handle != null, 'Unable to lookup callback.');
     _invokeMethod('startHeadlessService', {
-      'handleKey': <dynamic>[handle.toRawHandle()]
+      'handleKey': <dynamic>[handle?.toRawHandle()]
     });
 
     return;
@@ -367,13 +364,10 @@ class AudioPlayer {
   Future<bool> monitorNotificationStateChanges(
     void Function(AudioPlayerState value) callback,
   ) async {
-    if (callback == null) {
-      throw ArgumentError.notNull('callback');
-    }
-    final CallbackHandle handle = PluginUtilities.getCallbackHandle(callback);
-
+    final CallbackHandle? handle = PluginUtilities.getCallbackHandle(callback);
+    assert(handle != null, 'Unable to lookup callback.');
     await _invokeMethod('monitorNotificationStateChanges', {
-      'handleMonitorKey': <dynamic>[handle.toRawHandle()]
+      'handleMonitorKey': <dynamic>[handle?.toRawHandle()]
     });
 
     return true;
@@ -387,29 +381,26 @@ class AudioPlayer {
   /// respectSilence and stayAwake are not implemented on macOS.
   Future<int> play(
     String url, {
-    bool isLocal,
+    bool? isLocal,
     double volume = 1.0,
     // position must be null by default to be compatible with radio streams
-    Duration position,
+    Duration? position,
     bool respectSilence = false,
     bool stayAwake = false,
     bool duckAudio = false,
     bool recordingActive = false,
   }) async {
     isLocal ??= isLocalUrl(url);
-    volume ??= 1.0;
-    respectSilence ??= false;
-    stayAwake ??= false;
 
     final int result = await _invokeMethod('play', {
       'url': url,
       'isLocal': isLocal,
       'volume': volume,
       'position': position?.inMilliseconds,
-      'respectSilence': respectSilence ?? false,
-      'stayAwake': stayAwake ?? false,
-      'duckAudio': duckAudio ?? false,
-      'recordingActive': recordingActive ?? false,
+      'respectSilence': respectSilence,
+      'stayAwake': stayAwake,
+      'duckAudio': duckAudio,
+      'recordingActive': recordingActive,
     });
 
     if (result == 1) {
@@ -426,16 +417,12 @@ class AudioPlayer {
     Uint8List bytes, {
     double volume = 1.0,
     // position must be null by default to be compatible with radio streams
-    Duration position,
+    Duration? position,
     bool respectSilence = false,
     bool stayAwake = false,
     bool duckAudio = false,
     bool recordingActive = false,
   }) async {
-    volume ??= 1.0;
-    respectSilence ??= false;
-    stayAwake ??= false;
-
     if (!Platform.isAndroid) {
       throw PlatformException(
         code: 'Not supported',
@@ -551,10 +538,10 @@ class AudioPlayer {
   ///
   /// Specify atleast title
   Future<dynamic> setNotification({
-    String title,
-    String albumTitle,
-    String artist,
-    String imageUrl,
+    String title = '',
+    String albumTitle = '',
+    String artist = '',
+    String imageUrl = '',
     Duration forwardSkipInterval = Duration.zero,
     Duration backwardSkipInterval = Duration.zero,
     Duration duration = Duration.zero,
@@ -563,10 +550,10 @@ class AudioPlayer {
     bool hasNextTrack = false,
   }) {
     return _invokeMethod('setNotification', {
-      'title': title ?? '',
-      'albumTitle': albumTitle ?? '',
-      'artist': artist ?? '',
-      'imageUrl': imageUrl ?? '',
+      'title': title,
+      'albumTitle': albumTitle,
+      'artist': artist,
+      'imageUrl': imageUrl,
       'forwardSkipInterval': forwardSkipInterval.inSeconds,
       'backwardSkipInterval': backwardSkipInterval.inSeconds,
       'duration': duration.inSeconds,
@@ -623,7 +610,7 @@ class AudioPlayer {
     _log('_platformCallHandler call ${call.method} $callArgs');
 
     final playerId = callArgs['playerId'] as String;
-    final AudioPlayer player = players[playerId];
+    final AudioPlayer? player = players[playerId];
 
     if (!kReleaseMode && Platform.isAndroid && player == null) {
       final oldPlayer = AudioPlayer(playerId: playerId);
